put mimi_model into a shared header
